+++ CS342Application/Models/Member.swift
@@ -7,13 +7,8 @@
 //
 
 enum Member: String, Identifiable, CaseIterable {
-    case oliverAalami = "Prof. Oliver Aalami, MD"
     case paulSchmiedmayer = "Paul Schmiedmayer, Ph.D."
-
     case jenniferXu = "Jennifer Xu"
-    case vishnuRavi = "Vishnu Ravi, MD"
-    case rajPabari = "Raj Pabari"
-    case ashleyGriffin = "Ashley Griffin, Ph.D."
     
     
     static var allMembers: [Role: [Member]] {
@@ -41,13 +36,9 @@ enum Member: String, Identifiable, CaseIterable {
     
     var role: Role {
         switch self {
-        case .paulSchmiedmayer, .vishnuRavi, .ashleyGriffin:
+        case .paulSchmiedmayer:
             return .instructor
         case .jenniferXu:
-            return .student
-        case .oliverAalami:
-            return .lead
-        case .rajPabari:
             return .student
         }
     }
