+++ CS342Application/Views/Member+Views.swift
@@ -6,11 +6,8 @@
 // SPDX-License-Identifier: MIT
 //
 
-import AshleyGriffin
-import OliverAalami
 import PaulSchmiedmayer
 import JenniferXu
-import RajPabari
 import SwiftUI
 
 
@@ -23,14 +20,6 @@ extension Member {
                 PaulSchmiedmayer(navigationPath: navigationPath)
             case .jenniferXu:
                 JenniferXu(navigationPath: navigationPath)
-            case .vishnuRavi:
-                Text("Vishnu 👋")
-            case .oliverAalami:
-                OliverAalami(navigationPath: navigationPath)
-            case .rajPabari:
-                RajPabari()
-            case .ashleyGriffin:
-                AshleyGriffin(navigationPath: navigationPath)
             }
         }
         return memberView
