--- CS342MemberViews/Sources/AshleyGriffin/Resources/AshleyGriffinBio.md
@@ -1,7 +1,0 @@
-Dr. Ashley Griffin is a medical informatics postdoctoral fellow in the Department of Health Policy at Stanford and the VA Palo Alto Health Care System. Her research focuses on the use of digital health technologies and patient-generated health data to facilitate shared decision-making and inform interoperability initiatives.
-
-**Hobbies**
-- Cross-stitch 🪡
-- Piano 🎹
-- Yoga 🧘‍♀️
-- Listening to Taylor Swift 🕛
