--- CS342MemberViews/Sources/AshleyGriffin/Resources/AshleyGriffinBio.md.license
@@ -1,7 +1,0 @@
-
-This source file is part of the CS342 2023 Application project
-
-SPDX-FileCopyrightText: 2023 Stanford University
-
-SPDX-License-Identifier: MIT
-
