--- CS342MemberViews/Sources/OliverAalami/OliverDetailDestinations.swift
@@ -1,16 +1,0 @@
-//
-// This source file is part of the CS342 2023 Application project
-//
-// SPDX-FileCopyrightText: 2023 Stanford University
-//
-// SPDX-License-Identifier: MIT
-//
-
-enum OliverDetailDestinations: String, Identifiable {
-    case biodesignOffice
-    
-    
-    var id: String {
-        rawValue
-    }
-}
