--- CS342MemberViews/Sources/OliverAalami/Resources/OliverAalamiBio.md
@@ -1,13 +1,0 @@
-Dr. Aalami is a Clinical Professor of Vascular & Endovascular Surgery at Stanford University and the Palo Alto VA and serves as the Director of Stanford’s Biodesign for Digital Health. He is the course director for Biodesign for Digital Health, Building for Digital Health and co-founder of the open source project, CardinalKit, developed to support sensor-based mobile research projects. His primary research focuses on clinically validating the sensors in smartphones and smartwatches in patients with cardiovascular disease and translation of digital health solutions.  
-
-**Research Interests**
-
-- Data Sciences
-- Research Methods
-- Standards
-- Technology and Education
-
-
-**Current Research and Scholarly Interests**
-
-We launched a national precision medicine PAD trial called, [VascTrac](http://vasctrac.stanford.edu/). This trial is mobile phone based and leverages Apple's ResearchKit Platform to monitor a patient's activity both pre- and post-intervention. We are validating mobile phone surveillance for PAD patients and are currently enrolling.
