--- CS342MemberViews/Sources/RajPabari/Resources/RajPabari.jpeg.license
@@ -1,6 +1,0 @@
-
-This source file is part of the CS342 2023 Application project
-
-SPDX-FileCopyrightText: 2023 Stanford University
-
-SPDX-License-Identifier: MIT
