--- CS342MemberViews/Sources/RajPabari/Resources/RajPabariBio.md
@@ -1,5 +1,0 @@
-Raj Pabari is a a driven, inquisitive, outgoing self-starter with a passion for learning and inventing. His background combines experience in startups, projects, and corporate environments.
-
-Pabari finds unique solutions to real-world problems and has created unique solutions in various fields including renewable energy, peer-to-peer tutoring, biotechnology, online education, and water leak detectors. He innovates at the intersection of technology, business, and impact.
-
-Learn more at his website, rajpabari.com.
